tests/submission/Infraware/Forms/contents/Forms/meter_labels.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>meter_labels</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<form method="post"
		  enctype="application/x-www-form-urlencoded"
		  action="" 
		  id="input_form">
	 <p><label>Full name<label>(name):<meter id='meter_id1'></meter></label></label></p>
	 <p><label>Age:<meter id='meter_id2'></meter></label></p>
	</form>
	<script>

		var meter1 = document.getElementById("meter_id1");
		var meter2 = document.getElementById("meter_id2");

		if (typeof(meter1.labels) == "object") {
			if (meter1.labels.length == 2 && meter2.labels.length == 1) {
				test(function() {
					assert_true(true, "labels attribute is correct.");
				});
			} else {
				test(function() {
					assert_unreached("labels attribute is not correct.");
				});
			}
		} else {
			test(function() {
				assert_unreached("labels attribute is not exist.");
			});
		}

	</script>

 </body>
</html>
